api-gateway-app/app/requirements.txt
fastapi
uvicorn
httpx
aio-pika
aiohttp
python-dotenv
